ocr.h nativeness
- moved ALL of the functions from ocrfunctions.cpp to native.cpp, thats how windows-dependent this file was.
closes #28
