+++ src/native/native.hpp
@@ -12,6 +12,15 @@ void unregisterNativeHotkeys(std::map<int, KeyBind>& keyBindings);
 int getNativeScreenResolution();
 std::string VirtualKeyCodeToString(UCHAR virtualKey);
 
+HBITMAP takeScreenshot(int imageWidth, int imageHeight, int offsetX, int offsetY);
+HBITMAP takeScreenshot(int imageWidth, int imageHeight, Point);
+BOOL SaveHBITMAPToFile(HBITMAP hBitmap, LPCTSTR lpszFileName);
+void saveScreenshotToClipboard(HBITMAP bitmap);
+HGLOBAL loadStringToMemory(std::string& text);
+void saveItemsToClipboard(AppState& state, std::vector<Item>& items);
+
+
+
 #endif
 
 #if __linux__
@@ -22,6 +31,15 @@ void listenAndHandleEvents();
 void nativeWindowCustomization();
 void registerNativeHotkeys();
 void unregisterNativeHotkeys();
+
+
+
+void takeScreenshot();
+void takeScreenshot();
+void SaveHBITMAPToFile();
+void saveScreenshotToClipboard();
+void loadStringToMemory();
+void saveItemsToClipboard();
 
 
 #endif
